[quant][refactor] Factor out MHA code dup from nn and nn.quantizable
Quantizable MHA shares some code with its FP equivalent.
This factors out common parts.

Test Plan:

Explicit testing is not available, as there is no functional tests for the nn.MHA.
However, the functional equivalence of this PR was tested across commits by comparing the numerics of the pre-change and post-change MHA. Similar was done for the quantizable counterpart.

ghstack-source-id: 5f49abe40c026d79e0c8e8bc3872da23a78de397
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51169